Styling and date format

+++ src/scripts/news.js
@@ -41,9 +41,6 @@ const newsArticles = {
         // Create doc frag to hold each article item
         let newsDocFrag = document.createDocumentFragment();
 
-        // let sessionUser = sessionStorage.getItem("User");
-        // console.log("Session User:", sessionUser);
-
         // Create empty array
         let articleArray = [];
 
@@ -51,12 +48,10 @@ const newsArticles = {
         data.getFriends()                           // gets all connections filtered by sessionUser
         .then(allFriends => {
             allFriends.forEach(friend => {
-                // console.log("Friend ID:", friend.otherId);
                 let friendID = friend.otherId;
                 data.newsDataFriends(friendID)
                 .then(allFriendArticles => {
                     allFriendArticles.forEach(article => {
-                        // console.log("Friend ID:", friendID, "Article Title:", friendArticle.title);
                         articleArray.push(article);
                     })
                 })
@@ -66,23 +61,26 @@ const newsArticles = {
             data.newsData()
             .then(allUserArticles => {
                 allUserArticles.forEach(article => {
-                    // console.log("Session User's Article Title:", articleObj.title);
                     articleArray.push(article);
                 })
             })
             .then(function() {
-                // console.log("Final array", articleArray);
 
                 articleArray.sort(function(x, y){
                     return x.timestamp - y.timestamp;
                 })
 
-                // stuff goes here
                 articleArray.forEach(article => {
-                    // console.log(article.title);
+
+                        let sessionUser = sessionStorage.getItem("User");
 
                         let articleSection = document.createElement("div");
                         articleSection.setAttribute("class", "article_section");
+
+                        // Conditional for article styling
+                        if (article.userId !== Number(sessionUser)) {
+                            articleSection.classList.add("friend_article");
+                        }
 
                         let articleInfo = document.createElement("div");
                         articleInfo.setAttribute("class", "news_info");
@@ -103,59 +101,39 @@ const newsArticles = {
 
                         let newsAuthor = document.createElement("div");
                         newsAuthor.setAttribute("class", "news_author");
-                        newsAuthor.innerHTML = `Posted by ${article.user.name} on ${article.date}`;
-                        // console.log(article.date);
+
+                        if (Number(sessionUser) === article.userId) {
+                            newsAuthor.innerHTML = `You posted on ${article.date}`;
+                        }
+                        else {
+                            newsAuthor.innerHTML = `${article.user.name} posted on ${article.date}`;
+                        }
 
                         // Add author & delete button to etc
                         newsEtc.appendChild(newsAuthor);
 
-                        // Add conditional
+                        // Add conditional for "Delete" button
                         // If userId = current user, show "Delete" button
-                        // let sessionUser = sessionStorage.getItem("User");
-                        // console.log("Session:", sessionUser, " / Author:", article.userId)
-                        // if (sessionUser === article.userId) {
-                        //     let deleteArticleBtn = document.createElement("button");
-                        //     deleteArticleBtn.setAttribute("class", "article_delete_btn");
-                        //     deleteArticleBtn.textContent = "Delete";
+                        if (Number(sessionUser) === article.userId) {
+                            let deleteArticleBtn = document.createElement("button");
+                            deleteArticleBtn.setAttribute("class", "article_delete_btn");
+                            deleteArticleBtn.textContent = "Delete";
 
-                        //     deleteArticleBtn.addEventListener("click", () => {
-                        //         let articleId = article.id;
-                        //         data.deleteNewsData(articleId)
-                        //         .then(response => {
-                        //             newsArticles.buildArticles()
-                        //         })
-                        //     })
-
-                        //     newsEtc.appendChild(deleteArticleBtn);
-                        // }
-
-                        let deleteArticleBtn = document.createElement("button");
-                        deleteArticleBtn.setAttribute("class", "article_delete_btn");
-                        deleteArticleBtn.textContent = "Delete";
-
-                        deleteArticleBtn.addEventListener("click", () => {
-                            let articleId = article.id;
-                            data.deleteNewsData(articleId)
-                            .then(response => {
-                                newsArticles.buildArticles()
+                            deleteArticleBtn.addEventListener("click", () => {
+                                let articleId = article.id;
+                                data.deleteNewsData(articleId)
+                                .then(() => {
+                                    newsArticles.buildArticles()
+                                })
                             })
-                        })
-                        newsEtc.appendChild(deleteArticleBtn);
 
-                        // let sessionUser = sessionStorage.getItem("User");
-                        // console.log("Session:", sessionUser, " / Author:", article.userId)
-                        // if (article.userId !== sessionUser) {
-                        //     newsEtc.removeChild(deleteArticleBtn);
-                        // }
+                            newsEtc.appendChild(deleteArticleBtn);
+                        }
 
                         // Add title, synopsis, and URL to article info
                         articleInfo.appendChild(newsTitle);
                         articleInfo.appendChild(newsSynopsis);
                         articleInfo.appendChild(newsURL);
-
-                        // // Add author & delete button to etc
-                        // newsEtc.appendChild(newsAuthor);
-                        // newsEtc.appendChild(deleteArticleBtn);
 
                         // Add article info & etc to article section
                         articleSection.appendChild(articleInfo);
